Update dependencies and replace `Float` with `f32`.
`Float` has been removed from `amethyst`.

Issue #146

+++ crate/game_input_stdio/src/control_args.rs
@@ -12,7 +12,7 @@ pub enum ControlArgs {
         /// Axis to control, "x" or "z".
         axis: Axis,
         /// Value to use for the axis input.
-        value: f64,
+        value: f32,
     },
     /// Axis input.
     Action {
